+++ src/application/common/hash/hashes/argon2id.rs
@@ -8,7 +8,7 @@ build_hash! { ARGON2ID (reader, salt_opt) {
         Some(val) => val,
         None => {
             let mut rng = rand::thread_rng();
-            let mut bytes = [0u8; argon2::password_hash::Salt::recommended_len()];
+            let mut bytes = [0u8; argon2::password_hash::Salt::RECOMMENDED_LENGTH];
             rng.fill_bytes(&mut bytes);
             String::from(argon2::password_hash::SaltString::b64_encode(&bytes).expect("WhiteBeam: Salt string invariant violated").as_str())
         }
@@ -16,5 +16,5 @@ build_hash! { ARGON2ID (reader, salt_opt) {
     // Argon2 with default params (Argon2id v19)
     let argon2 = argon2::Argon2::default();
     // Hash password to PHC string ($argon2id$v=19$...)
-    argon2.hash_password_simple(password.as_bytes(), salt.as_ref()).unwrap().to_string()
+    argon2.hash_password_simple::<String>(password.as_bytes(), &salt).unwrap().to_string()
 }}